feat: Restructure IAM routing from /admin to /admin/iam
- Create new /admin/iam route structure for Identity Access Manager
- Update menu configuration to point IAM to /admin/iam path
- Update Users menu item to link to /admin/iam?tab=users
- Add redirect page at /admin that forwards to /admin/iam
- Fix all API endpoints to use consistent port 3000 instead of 5000
- Update activePage parameter from 'admin' to 'iam' for proper navigation
- Maintain IAM branding and functionality while improving URL structure
